+++ changelog.md
@@ -7,6 +7,7 @@ This version is not released yet and is under active development.
 ```
 
 - Add colorization of options, choices and metavars in help screens.
+- Add `@version_option` decorator to colorize ``--version`` option output.
 
 ## [0.0.1 (2021-10-18)](https://github.com/kdeldycke/click-extra/compare/88b81e...v0.0.1)